requirements.txt
dataclasses==0.6
MIDIUtil==1.2.1
mido-fix==1.2.12
musicpy==6.91
numpy==2.0.0
pyFluidSynth==1.3.3
pygame-ce==2.5.0
pygame_gui==0.6.12
python-i18n==0.3.9
